disparo de proyectiles

+++ skeleton/Particle.cpp
@@ -17,8 +17,9 @@ Particle::~Particle()
 void Particle::integrate(double t)
 {
 	//MRU
+	//if (inverse_mass <= 0.0f) return;
 	pose = physx::PxTransform(pose.p.x + vel.x * t, pose.p.y + vel.y * t, pose.p.z + vel.z * t);
 	vel += acc * t;
-	vel *= pow(damping, t);
+	vel *= powf(damping, t);
 
 }
